src/indexer/requirements.txt
opensearch-py==2.4.2
boto3-stubs[s3]==1.34.42
